+++ binaries/tags/src/Commands/DockerHubTagsCommand.ts
@@ -1,23 +1,28 @@
 import { Arguments, CommandModule } from 'yargs'
 
 import { Builder } from './Builder'
-import { DockerTags } from '../DockerHubTags'
+import { DockerHubTags } from '../DockerHubTags'
 
 export interface DockerHubOptions extends Arguments {
   repository: string
-  tag: string
+  tag?: string
 }
 
-export class DockerHubTags implements CommandModule<{}, DockerHubOptions> {
-  command = '$0 <repository> <tag>'
+export class DockerHubTagsCommand implements CommandModule<{}, DockerHubOptions> {
+  command = '$0 <repository> [tag]'
 
   builder: Builder = {}
 
   handler = async (args: DockerHubOptions) => {
-    const search = new DockerTags(args.repository)
+    const search = new DockerHubTags(args.repository)
     const searcher = await search.tags()
-    console.log(`v${searcher.latest(args.tag)}`)
+
+    if (args.tag) {
+      console.log(searcher.latest(args.tag))
+    } else {
+      console.log(searcher.enumerate().map(tag => tag.version))
+    }
   }
 }
 
-export const DockerHubTagsCommand = new DockerHubTags()
+export const TagsCommand = new DockerHubTagsCommand()