+++ binaries/tags/src/Tag.ts
@@ -1,4 +1,4 @@
 export interface Tag {
-  layer: string
   name: string
+  version: string
 }