+++ binaries/tags/src/TagInfo.ts
@@ -1,0 +1,4 @@
+export interface TagInfo {
+  layer: string
+  name: string
+}