+++ binaries/tags/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.5](https://github.com/nativecode-dev/cli-tools/compare/@nofrills/tags@0.1.4...@nofrills/tags@0.1.5) (2020-02-12)
+
+
+### Bug Fixes
+
+* remove semver import ([1d58961](https://github.com/nativecode-dev/cli-tools/commit/1d58961e4850ce06eaec5bf3b5406e6e6d814b29))
+* **replace semver:** replaced semver with compare-versions ([1ed4378](https://github.com/nativecode-dev/cli-tools/commit/1ed4378457a1c1688a32101ba6c050025de74f75))
+
+
+
+
+
 ## [0.1.4](https://github.com/nativecode-dev/cli-tools/compare/@nofrills/tags@0.1.3...@nofrills/tags@0.1.4) (2020-02-12)
 
 **Note:** Version bump only for package @nofrills/tags